Extract test of MockCurrentUser into new module
