+++ tests/no_version/test_auth_api.py
@@ -387,23 +387,3 @@ async def test_delete_user(registered_client, registered_superuser_client):
         f"{PREFIX}/users/THIS-IS-NOT-AN-ID"
     )
     assert res.status_code == 404
-
-
-@pytest.mark.parametrize("cache_dir", ("/some/path", None))
-@pytest.mark.parametrize("username", ("my_username", None))
-@pytest.mark.parametrize("slurm_user", ("test01", None))
-async def test_MockCurrentUser_fixture(
-    MockCurrentUser,
-    cache_dir,
-    username,
-    slurm_user,
-):
-
-    user_kwargs = dict(
-        cache_dir=cache_dir, username=username, slurm_user=slurm_user
-    )
-    async with MockCurrentUser(user_kwargs=user_kwargs) as user:
-        debug(user)
-        assert user.cache_dir == cache_dir
-        assert user.username == username
-        assert user.slurm_user == slurm_user
